requirements.txt
asgiref==3.11.0
black==26.1.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coverage==7.10.6
Django==5.2.11
django-allauth==64.2.0
django-widget-tweaks==1.5.1
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
pillow==12.1.1
platformdirs==4.5.1
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.12.0
python-dotenv==1.2.2
pytokens==0.4.1
PyYAML==6.0.3
requests==2.32.5
responses==0.26.0
smartystreets_python_sdk==5.5.0
SQLAlchemy==2.0.44
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.5.0
Werkzeug==3.1.3
PyYAML==6.0.3
